# assets/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Category & Product CRUD API",
    "description": "RESTful API untuk manajemen kategori dan produk menggunakan Go vanilla (tanpa framework) dengan PostgreSQL (Neon).\n\n## Features\n- CRUD lengkap untuk kategori dan produk\n- Service Layer architecture\n- Dependency Injection pattern\n- Validasi input\n- Error handling yang proper\n- CORS enabled",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "description": "Mengambil semua kategori yang tersedia",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category",
        "description": "Membuat kategori baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              },
              "example": {
                "name": "Electronics",
                "description": "Electronic devices and gadgets"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category by ID",
        "description": "Mengambil detail satu kategori berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update category",
        "description": "Mengupdate kategori yang sudah ada",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              },
              "example": {
                "name": "Updated Electronics",
                "description": "Updated description"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete category",
        "description": "Menghapus kategori berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Category deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "Mengambil semua produk yang tersedia. Bisa difilter berdasarkan nama produk.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "description": "Search products by name (case-insensitive, partial match)",
            "schema": {
              "type": "string",
              "example": "laptop"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a new product",
        "description": "Membuat produk baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              },
              "example": {
                "name": "Laptop Gaming",
                "price": 18500000,
                "stock": 15,
                "category_id": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by ID",
        "description": "Mengambil detail satu produk berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product",
        "description": "Mengupdate produk yang sudah ada",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              },
              "example": {
                "name": "Laptop Gaming Pro",
                "price": 22000000,
                "stock": 10,
                "category_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product",
        "description": "Menghapus produk berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkout": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create a transaction (Checkout)",
        "description": "Membuat transaksi baru dengan item-item produk. Stok produk akan otomatis dikurangi.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              },
              "example": {
                "items": [
                  {
                    "product_id": 1,
                    "quantity": 2
                  },
                  {
                    "product_id": 2,
                    "quantity": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transactions/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get transaction by ID",
        "description": "Mengambil detail satu transaksi berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Transaction ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/report": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get transaction report",
        "description": "Mengambil laporan transaksi berdasarkan rentang tanggal",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "description": "Start date (format: YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-01-01"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "description": "End date (format: YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-02-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid date format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Category ID (auto-generated)",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "description": "Category description",
            "example": "Electronic devices and gadgets"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2026-01-25T13:38:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2026-01-25T13:38:00Z"
          }
        }
      },
      "CreateCategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name (required)",
            "example": "Electronics",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "Category description (optional)",
            "example": "Electronic devices and gadgets"
          }
        }
      },
      "UpdateCategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name (required)",
            "example": "Updated Electronics",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "Category description (optional)",
            "example": "Updated description"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Error message detail",
            "example": "name is required"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Product ID (auto-generated)",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Laptop Gaming"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price",
            "example": 18500000
          },
          "stock": {
            "type": "integer",
            "description": "Product stock quantity",
            "example": 15
          },
          "category_id": {
            "type": "integer",
            "description": "ID of the category",
            "example": 1
          },
          "category_name": {
            "type": "string",
            "description": "Name of the category",
            "example": "Electronics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2026-02-01T21:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2026-02-01T21:30:00Z"
          }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": [
          "name",
          "price",
          "stock"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name (required)",
            "example": "Laptop Gaming",
            "minLength": 1
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price (required)",
            "example": 18500000,
            "minimum": 0
          },
          "stock": {
            "type": "integer",
            "description": "Product stock quantity (required)",
            "example": 15,
            "minimum": 0
          },
          "category_id": {
            "type": "integer",
            "description": "ID of the category (required)",
            "example": 1,
            "minimum": 1
          }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "required": [
          "name",
          "price",
          "stock",
          "category_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name (required)",
            "example": "Laptop Gaming Pro",
            "minLength": 1
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price (required)",
            "example": 22000000,
            "minimum": 0
          },
          "stock": {
            "type": "integer",
            "description": "Product stock quantity (required)",
            "example": 10,
            "minimum": 0
          },
          "category_id": {
            "type": "integer",
            "description": "ID of the category (required)",
            "example": 1,
            "minimum": 1
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Transaction ID (auto-generated)",
            "example": 1
          },
          "total_amount": {
            "type": "integer",
            "description": "Total amount in cents",
            "example": 3700000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Transaction timestamp",
            "example": "2026-02-08T13:30:00Z"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDetail"
            }
          }
        }
      },
      "TransactionDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Detail ID (auto-generated)",
            "example": 1
          },
          "transaction_id": {
            "type": "integer",
            "description": "Transaction ID",
            "example": 1
          },
          "product_id": {
            "type": "integer",
            "description": "Product ID",
            "example": 1
          },
          "product_name": {
            "type": "string",
            "description": "Product name",
            "example": "Laptop Gaming"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity purchased",
            "example": 2
          },
          "subtotal": {
            "type": "integer",
            "description": "Subtotal in cents",
            "example": 3700000
          }
        }
      },
      "CheckoutItem": {
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "properties": {
          "product_id": {
            "type": "integer",
            "description": "Product ID",
            "example": 1,
            "minimum": 1
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity to purchase",
            "example": 2,
            "minimum": 1
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "Array of items to checkout",
            "items": {
              "$ref": "#/components/schemas/CheckoutItem"
            },
            "minItems": 1
          }
        }
      },
      "ReportResponse": {
        "type": "object",
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "Start date of report",
            "example": "2026-01-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "End date of report",
            "example": "2026-02-01"
          },
          "total_transactions": {
            "type": "integer",
            "description": "Total number of transactions in period",
            "example": 5
          },
          "total_revenue": {
            "type": "integer",
            "description": "Total revenue in cents",
            "example": 18500000
          },
          "transactions": {
            "type": "array",
            "description": "List of transactions in period",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      }
    }
  }
}